Moved most parts of the DescriptorAbstract and TagDescriptor into traits
