+++ src/phpDocumentor/Descriptor/TagDescriptor.php
@@ -25,14 +25,9 @@ use phpDocumentor\Reflection\DocBlock\Description;
  */
 class TagDescriptor implements Filterable
 {
-    /** @var string $name Name of the tag. */
-    protected $name;
-
-    /** @var DescriptionDescriptor $description A description line for this tag */
-    protected $description;
-
-    /** @var Collection<Validation\Error> A collection of errors found during filtering. */
-    protected $errors;
+    use Traits\HasName;
+    use Traits\HasDescription;
+    use Traits\HasErrors;
 
     /**
      * Initializes the tag by setting the name and errors,
@@ -40,64 +35,11 @@ class TagDescriptor implements Filterable
     public function __construct(string $name, ?DescriptionDescriptor $description = null)
     {
         $this->setName($name);
-        $this->errors = Collection::fromClassString(Validation\Error::class);
         $this->setDescription($description);
-    }
-
-    /**
-     * Sets the name for this tag.
-     */
-    protected function setName(string $name): void
-    {
-        $this->name = $name;
-    }
-
-    /**
-     * Returns the name for this tag.
-     */
-    public function getName(): string
-    {
-        return $this->name;
-    }
-
-    /**
-     * Sets a description for this tab instance.
-     */
-    public function setDescription(?DescriptionDescriptor $description): void
-    {
-        $this->description = $description ?? new DescriptionDescriptor(new Description(''), []);
-    }
-
-    /**
-     * Returns the description for this tag,
-     */
-    public function getDescription(): DescriptionDescriptor
-    {
-        return $this->description;
     }
 
     public function __toString(): string
     {
         return (string) $this->description;
-    }
-
-    /**
-     * Sets a list of errors found on the usage of this tag.
-     *
-     * @param Collection<Validation\Error> $errors
-     */
-    public function setErrors(Collection $errors): void
-    {
-        $this->errors = $errors;
-    }
-
-    /**
-     * Returns all errors associated with this tag.
-     *
-     * @return Collection<Validation\Error>
-     */
-    public function getErrors(): Collection
-    {
-        return $this->errors;
     }
 }
