+++ src/phpDocumentor/Descriptor/Traits/HasErrors.php
@@ -1,0 +1,40 @@
+<?php
+
+declare(strict_types=1);
+
+namespace phpDocumentor\Descriptor\Traits;
+
+use phpDocumentor\Descriptor\Collection;
+use phpDocumentor\Descriptor\Validation\Error;
+
+trait HasErrors
+{
+    /** @var Collection<Error> A collection of errors found during filtering. */
+    protected Collection $errors;
+
+    /**
+     * Sets a list of all errors associated with this element.
+     *
+     * @internal should not be called by any other class than the assemblers
+     *
+     * @param Collection<Error> $errors
+     */
+    public function setErrors(Collection $errors): void
+    {
+        $this->errors = $errors;
+    }
+
+    /**
+     * Returns all errors associated with this tag.
+     *
+     * @return Collection<Error>
+     */
+    public function getErrors(): Collection
+    {
+        if (!isset($this->errors)) {
+            $this->errors = Collection::fromClassString(Error::class);
+        }
+
+        return $this->errors;
+    }
+}
